+++ pages/footer.php
@@ -10,6 +10,6 @@
 <script>
 $('#footer_release').click(function() {
 	$('#release_notes').modal('show');
-	console.log('rel');
+	//console.log('rel');
 });
 </script>