+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New:
 ----------------------
-- Bug fix release
+- Introducing HTML Templates. Now you can use your own format to template IFRA certifications
+- Fix the database restore backup process via UI
+- Fix a bug preventing formula pdf export in Chrome